+++ yt/server/scheduler/merge_controller.cpp
@@ -157,7 +157,7 @@ protected:
 
         virtual int GetChunkListCountPerJob() const override
         {
-            return static_cast<int>(Controller->OutputTables.size());
+            return Controller->OutputTables.size();
         }
 
         virtual TNullable<i64> GetJobDataSizeThreshold() const override
